Move into component dir
